// src/template/base/models/message.model.ts
export enum ConsoleMessage {
    TITLE = 'TPM',
    BANNER = 'Welcome to template package mananger \nbuild your application with drag and design',
    ERROR = 'ERROR: ',
    SUCCESS = 'SUCCESS: ',
    INFO = 'INFO: ',
    GENERATE = 'GENERATE: ',
    CREATE = 'CREATE: ',
    UPDATE = 'UPDATE: ',
    WARN ='WARN',
    START_GENERATING = 'Start generating a component...',
}